Moves `ErrCodeEquals` to `tfawserr`
