+++ errors.go
@@ -4,7 +4,6 @@
 package awsbase
 
 import (
-	smithy "github.com/aws/smithy-go"
 	"github.com/hashicorp/aws-sdk-go-base/v2/internal/config"
 	"github.com/hashicorp/aws-sdk-go-base/v2/internal/errs"
 )
@@ -23,20 +22,4 @@ type NoValidCredentialSourcesError = config.NoValidCredentialSourcesError
 // IsNoValidCredentialSourcesError returns true if the error contains the NoValidCredentialSourcesError type.
 func IsNoValidCredentialSourcesError(err error) bool {
 	return errs.IsA[NoValidCredentialSourcesError](err)
-}
-
-// AWS SDK for Go v2 variants of helpers in v2/awsv1shim/tfawserr.
-
-// ErrCodeEquals returns true if the error matches all these conditions:
-//   - err is of type smithy.APIError
-//   - Error.Code() equals one of the passed codes
-func ErrCodeEquals(err error, codes ...string) bool {
-	if apiErr, ok := errs.As[smithy.APIError](err); ok {
-		for _, code := range codes {
-			if apiErr.ErrorCode() == code {
-				return true
-			}
-		}
-	}
-	return false
 }
